pyproject.toml
[project]
name = "resumap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "huggingface-hub>=0.27.0",
    "llama-cpp-python>=0.3.0",
    "pydantic-settings>=2.7.0",
    "pymupdf>=1.25.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.18",
    "uvicorn[standard]>=0.32.0",
]
